Refactor folder structure and add boilerplate
