+++ app/src/main/java/com/urosmilosavljevic/foodapp/authentication/signup/presentation/SignUpScreen.kt
@@ -1,0 +1,14 @@
+package com.urosmilosavljevic.foodapp.authentication.signup.presentation
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+fun SignUpScreen() {
+}
+
+@Preview
+@Composable
+private fun SignUpScreenPreview() {
+    SignUpScreen()
+}
