+++ app/src/main/java/com/urosmilosavljevic/foodapp/authentication/signup/presentation/SignupForm.kt
@@ -1,0 +1,14 @@
+package com.urosmilosavljevic.foodapp.authentication.signup.presentation
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+fun SignupForm() {
+}
+
+@Preview
+@Composable
+fun PreviewSignupForm() {
+    SignupForm()
+}
