+++ app/src/main/java/com/urosmilosavljevic/foodapp/di/Modules.kt
@@ -1,8 +1,8 @@
 package com.urosmilosavljevic.foodapp.di
 
-import com.urosmilosavljevic.foodapp.authentication.domain.ValidateEmail
-import com.urosmilosavljevic.foodapp.authentication.domain.ValidatePassword
-import com.urosmilosavljevic.foodapp.authentication.presentation.LoginViewModel
+import com.urosmilosavljevic.foodapp.authentication.login.domain.ValidateEmail
+import com.urosmilosavljevic.foodapp.authentication.login.domain.ValidatePassword
+import com.urosmilosavljevic.foodapp.authentication.login.presentation.LoginViewModel
 import com.urosmilosavljevic.foodapp.onboarding.data.OnboardingPreferences
 import com.urosmilosavljevic.foodapp.onboarding.presentation.OnboardingViewModel
 import org.koin.core.module.dsl.viewModel
